Add partial endpoint tests for v1 API
I'm dubious about the value of this. On the one hand it caught an egregious
error before I deployed it, so it's worth having some of this. But on the other
hand, all the mocking kind of turns these tests into a pure tautology after a
while, and the part that ISN'T a tautology is already easily legible because the
route handlers in Express are so simple. Like, yeah, I'm technically testing
that Node didn't fuck up its implementation of times tables, but also it's
probably safe to just trust that two times four is eight.

So I don't think I'll add more of these for now unless I can figure out a
non-obnoxious way to switch all the session handling and passport stuff to use
the test database as well, and get some real-ass end-to-end stateful integration
tests working.