A different tack for securing CORS requests: save req.session.loginOrigin
The core here is, I want my endpoints to be able to do different (more
restrictive) things if a request is CORS. There's no way to detect a CORS
request from its headers in isolation (seems like a gap in the spec tbh), and I
really wanted to avoid hardcoding or configuring a list of allowed non-CORS
origin hostnames. This seemed like a good workaround, we'll see if it works.