Create form: fix initial selection range for sites with a root-level query, e.g. As The Crow Flies
Basically, it's always possible for the pathname to reoccur within the hostname
part of the URL, and it's guaranteed if the pathname is just '/'. Also, it's tempting
to use .origin, but that would theoretically bork on sites that use a username and
password in the URL (by the way, don't ever do that, thx). So that leaves .host.