(Broken-ish) Create (and expect!) dogear prefixes with leading "www." removed
This makes Eardogger ignore the leading `www.` on URLs, because I ran into some
sites that changed whether or not they use it over time, and they broke my
bookmarks. I think everyone recognizes that as just a tic by now, so we
shouldn't treat it as significant for matching purposes. (If you actually run a
domain that serves different content on `thing.co` and `www.thing.co`... you're
a degenerate, but, also sorry for the inconvenience.)

Anyway:

- Changes dogear creation to strip any leading `www.` from new prefixes.
- Changes dogear lookup to ignore leading "www."s when determining which
  prefixes match. (`www.thing.co/3` and `thing.co/3` will both match the
  `thing.co/` prefix.)

As currently implemented, this breaks existing dogears whose prefix includes a
`www.` -- they can't ever be matched, so you'd need to delete them and make new
bookmarks. The next commit has a fix for that.