Further refactor client JS
The way I had stuff before, it looked like there was more global state than
there actually was. Really, the only thing that touches the bookmarksList is the
refresh function, so I can just have it look for it, and the other three things
are also very tightly scoped so I can move the checks right next to them within
the whatever() call.