Use layout: false for fragments instead of a {{body}}-only layout
That {{body}} layout was adding entity escaping to my HTML when
I sent it over the wire, which, now that I look at main.hbs again, that was
supposed to be {{{body}}} (triple-curlies). But at any rate, it's totally
unnecessary and we can just pass false to get the desired effect.