Re-jigger db modules to double-check func signatures
Okay, this seems a little berserk at first, but essentially what I'm doing is
building myself a self-checking mock framework for all my database-hitting
functions.

Picking up this project after not touching it for a year and a half has
convinced me I need better test coverage in my API tests, and it might be nice
to have ANY tests for my UI features, and in order to do any of that I need WAY
more extensive DB mocking than I currently have. (It'd also be nice for the
pagination refactor I'm planning on dogears#list.)

Mocks are incredibly hazardous if they don't stay perfectly in sync with the
real functions, but it occurred to me that I could make Typescript do all the
heavy lifting for me, and get a shout whenever the args/returns of my mocks stop
matching reality. So that's the plan.